removed protocol buffers dependency
The synchronous interface currently requires protocol buffers, which are
a bit tricky to work with due to the need for generated code. Recently
there has been a breaking update in protocol buffers, which would tie
the version of ipfsspec to certain versions of protocol buffers, see
https://developers.google.com/protocol-buffers/docs/news/2022-05-06#python-updates
for more information about that.