remove multi-gateway functionality
In the past, the multi-gateway functionality often created more harm
than use. The functionality mainly existed, to provide a
zero-configuration experience for new users. Due to IPIP-280, this is
now less of a concern. Further, the new auto-config method encourages
new users to briefly interact with IPFS, maybe leading to a better
performing setup.